feat: add /usr/share as a config location
/usr/share/<application name> is becoming more and more common for
applications to store their default configuration in.
Either the application itself or distributions can use this location to
ship a config that users will not touch, and instead they are expected
to overwrite (parts of) the config in /etc/<application name>. This
means this config file can be safely updated with new required values
without having to worry about overwriting user-made change.

This is useful in general but especially in case of immutable
distributions who don't touch anything in /etc and /home